imx-boot: override the final bootloader binary
Since the U-Boot binary signing[1] was introduced, the final bootloader
binary, imx-boot, has become a symbolic link to imx-boot-tagged.

The imx-boot-tagged is imx-boot with a string of characters containing
the binary size to make it easier for UUU to find the end of it. This
string is appended to the end of the bootloader binary, modifying it.

HAB then recognizes that it was modified, causing the secure boot to
fail as a security measure.

This patch overrides the symbolic link to imx-boot-tagged, making
imx-boot to be the correct bootloader binary, signed by hab.

[1] Freescale/meta-freescale@92acf99a6124

Signed-off-by: Alifer Moraes <alifer.m@variscite.com>