classes/mx8-fuse-commands-helper: Create fuses command helper class
After the device successfully boots a signed image without generating any HAB
events, it is safe to **secure**, or **"close,"** the device. This is the final
step in the process. Once the fuse is blown, the chip will **only** load images
signed with the correct PKI tree.

Important Notes:
- This is a **One-Time Programmable (OTP) e-fuse**. Once written,
it **cannot be undone**, so make sure everything is correct before proceeding.

- If any of the previous steps were not performed correctly, the SOM
**will not boot** after this bit is written.

This class is based on the mx8_create_fuse_commands.sh script:
    - https://github.com/varigit-dev/meta-variscite-hab/blob/210d24670a7da37059ec35e76de6f0df953800e6/recipes-bsp/imx-mkimage/imx-boot-hab/mx8_create_fuse_commands.sh

Signed-off-by: Luan Rafael Carneiro <luan.r@variscite.com>