mx8-fuse-commands-helper: Add i.MX8 family support
Remove the overrides from do_compile and do_deply as this class
is valid for i.MX8 and i.MX8M family.

As the SRK fuse binary file name is differently for both families,
reflect this accordingly.

This also adds a warning to do_deploy in case the SRK fuse file
can not be created. The behavior is derived from create_fuse_cmds.

Signed-off-by: Alexander Herfurtner <alexander.h@variscite.com>